// apps/kennisbank-api/src/types/openapi.ts
export interface components {
    schemas: {
// ...
        VAC: {
            /** @description De vraag waar de VAC antwoord op geeft. */
            vraag: string;
            /**
             * @description De actuele status van de VAC.
             * @enum {string}
             */
            status: "actief" | "non-actief" | "te-verwijderen";
            /** @description Het antwoord op de bij 'vraag' gestelde vraag. */
            antwoord: string;
            /**
             * @description De doelgroep van de VAC.
             * @enum {string}
             */
            doelgroep: "eu-burger" | "eu-bedrijf" | "eu-burger-bedrijf";
            /**
             * Gerelateerde Afdelingen
             * @description Afdelingen aan wie de VAC gerelateerd is.
             */
            afdelingen?: {
                /** @description De unieke identifier van de afdeling. */
                afdelingId: string;
                /** @description De naam van de afdeling. */
                afdelingNaam: string;
            }[];
            /** @description Een toelichting bij de VAC ten behoeve van de medewerker van het contactcentrum. */
            toelichting?: string;
            /**
             * Trefwoorden
             * @description Trefwoorden die passen bij de VAC.
             */
            trefwoorden?: {
                /** @description Een trefwoord. */
                trefwoord: string;
            }[];
            /**
             * Gerelateerde VACs
             * @description VACs die gerelateerd zijn aan de VAC.
             */
            gerelateerdeVACs?: {
                /**
                 * Format: uri
                 * @description De URI van de gerelateerde VAC.
                 */
                VAC: string;
            }[];
            /**
             * Gerelateerde Producten
             * @description Producten die gerelateerd zijn aan de VAC.
             */
            gerelateerdeProducten?: {
                /**
                 * Format: uri
                 * @description De URI van het gerelateerde product.
                 */
                product: string;
            }[];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}